Fix feature for removing watches that never worked, now kinda works
The whole feature for removing watches was predicated us finding it,
but of course we're not going to find it if it's just been deleted.

Hack around that by keeping around the last list of files we watched,
and removing stuff from there that doesn't exist anymore.

That'll also have a race condition and we'll forget to remove things
if we end up doing something we don't get events for, but oh well.