Stop using File::Find::Rule and use find(1) again
Undo the change to use File::Find::Rule added in 4b556ae. See note in
64f0772 for a suggestion of this. I tested this and the program is
much faster (as expected) in the discovery phase and in adding
watches, which we do all the time.

I don't really see the point of not shelling out to POSIX utilities
when convenient when the whole module depends on a Linux-specific
feature. It's not like this is going to run on Win32 or Android.