Handle not ENOENT errors while adding watches
There's a race condition here that sometimes causes us to die and exit
the whole program, if a directory we're intending to watch is removed
between us running the find and running this code we'll hard die.

Just handle that gracefully and log it and continue.

Arguably this should be handling the case of:

    perl -Ilib bin/inotify-hookable --no-r -w $PWD/lib -w $PWD/bin -w $PWD/t -d

Where all the $PWD directories except t exist differently than the more
common case of some subdirectory of lib being created and then deleted,
i.e. maybe we should still die if a top-level watch goes missing.